refector the the non-latched frame buffer into `plain` module
