+++ examples/i2s_parallel.rs
@@ -64,7 +64,7 @@ use esp_hal::timer::timg::TimerGroup;
 use esp_hal_embassy::InterruptExecutor;
 use esp_hub75::framebuffer::compute_frame_count;
 use esp_hub75::framebuffer::compute_rows;
-use esp_hub75::framebuffer::DmaFrameBuffer;
+use esp_hub75::framebuffer::plain::DmaFrameBuffer;
 use esp_hub75::i2s_parallel::Hub75;
 use esp_hub75::Color;
 use esp_hub75::Hub75Pins16;
