+++ src/i2s_parallel.rs
@@ -9,7 +9,7 @@ use esp_hal::i2s::parallel::TxSixteenBits;
 use esp_hal::i2s::AnyI2s;
 use esp_hal::time::Rate;
 
-use crate::framebuffer::DmaFrameBuffer;
+use crate::framebuffer::plain::DmaFrameBuffer;
 use crate::Hub75Error;
 use crate::Hub75Pins16;
