+++ src/lcd_cam.rs
@@ -12,7 +12,7 @@ use esp_hal::lcd_cam::LcdCam;
 use esp_hal::peripherals::LCD_CAM;
 use esp_hal::time::Rate;
 
-use crate::framebuffer::DmaFrameBuffer;
+use crate::framebuffer::plain::DmaFrameBuffer;
 use crate::Hub75Error;
 use crate::Hub75Pins16;
